@@ -1,7 +1,6 @@
-import React, { createContext, useContext, useState } from "react";
+import React, { createContext, useContext, useState, useEffect } from "react";
 import PropTypes from "prop-types";
-import routes from "../constants/RouteConstants";
-import { useApiContext } from "./ApiContext";
+
 import { anonymousUser } from "../data/data";
 
 const UserContext = createContext(null);
@@ -12,24 +11,13 @@ export function useUserContext() {
 
 export function UserProvider({ children }) {
   const [user, setUser] = useState(anonymousUser);
-  const api = useApiContext();
 
-  const loggedIn = async () => {
-    try {
-      var userInfo = await api.apiGet(routes.userInfo);
-      console.log("LoggedIn:userInfo:", userInfo);
-      if (userInfo && userInfo.status === 200) {
-        setUser(userInfo.data);
-      } else {
-        setUser(anonymousUser);
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
+  useEffect(() => {
+    console.log("User Changed", user);
+  }, [user]);
 
   return (
-    <UserContext.Provider value={{ user, setUser, loggedIn }}>
+    <UserContext.Provider value={{ user, setUser }}>
       {children}
     </UserContext.Provider>
   );
